% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_tweedie.R
\name{tweedie}
\alias{tweedie}
\alias{rtweedie}
\alias{dtweedie}
\alias{ptweedie}
\alias{qtweedie}
\title{Tweedie Distribution Functions}
\usage{
rtweedie(n, mean = 1, dispersion = 1, power = 1.5)

dtweedie(x, mean = 1, dispersion = 1, power = 1.5, log = FALSE)

ptweedie(
  q,
  mean = 1,
  dispersion = 1,
  power = 1.5,
  lower.tail = TRUE,
  log.p = FALSE
)

qtweedie(
  p,
  mean = 1,
  dispersion = 1,
  power = 1.5,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}

\item{mean}{vector of means.}

\item{dispersion}{vector of dispersion parameters.}

\item{power}{vector of power parameters.}

\item{x, q}{vector of quantiles.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \le x]}; otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}
}
\value{
\code{dtweedie} gives the density, \code{ptweedie} gives the distribution
function, \code{qtweedie} gives the quantile function, and \code{rtweedie}
generates random samples.

The length of the result is determined by \code{n} for \code{rtweedie}, and is the
maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the
result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for
the Tweedie distribution with mean equal to \code{mean}, dispersion equal to
\code{dispersion}, and power equal to \code{power}.
}
\details{
If \code{mean}, \code{dispersion}, or \code{power} are not specified they assume the
default values of \code{1}, \code{1}, and \code{1.5}, respectively.

The Tweedie distribution used here follows the compound Poisson-Gamma
parameterisation with power parameter in \eqn{(1, 2)}. It has
\eqn{\mathbb{E}[X] = \mu} and
\eqn{\mathrm{Var}(X) = \phi\mu^p}, where \eqn{\mu} is \code{mean},
\eqn{\phi} is \code{dispersion}, and \eqn{p} is \code{power}.
}
\references{
Dunn, P. K., & Smyth, G. K. (2005). Series evaluation of Tweedie
exponential dispersion model densities. \emph{Statistics and Computing},
15(4), 267--280. \doi{10.1007/s11222-005-4070-y}.
}
