% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedieDistr_package.R
\docType{package}
\name{tweedieDistr-package}
\alias{tweedieDistr}
\alias{tweedieDistr-package}
\title{tweedieDistr: Tweedie Distribution}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides density, distribution function, quantile function, and random generation for the Tweedie distribution under the compound Poisson-Gamma parameterisation with power parameter in (1, 2). The density is evaluated using the series expansion of Dunn & Smyth (2005), implemented in C++ via Rcpp and RcppArmadillo for performance. A 'distributional'-compatible constructor is also provided for use in tidy modelling workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StefanoDamato/tweedieDistr}
  \item Report bugs at \url{https://github.com/StefanoDamato/tweedieDistr/issues}
}

}
\author{
\strong{Maintainer}: Stefano Damato \email{stefanodamato128@gmail.com}

}
\keyword{internal}
